Refactoring Predict + New operator ASSIGN_LEARNING_TAR
