+++ engine/register_model.cpp
@@ -18,7 +18,7 @@
 #include <configuration/default_config_manager.h>
 #include <configuration/model_configuration_manager.h>
 #include <jsoncpp/json/value.h>
-#include "include/ddl_operators.h"
+#include "include/ml_operators.h"
 #include <fstream>
 #include <jsoncpp/json/json.h>
